+++ src/native/nv_helpers_dx12/RootSignatureGenerator.cpp
@@ -62,7 +62,7 @@ namespace nv_helpers_dx12
         // The descriptor table descriptor ranges require a pointer to the descriptor ranges. Since new
         // ranges can be dynamically added in the vector, we separately store the index of the range set.
         // The actual address will be solved when generating the actual root signature
-        m_rangeLocations.push_back(static_cast<UINT>(m_ranges.size() - 1));
+        m_rangeLocations.push_back(m_ranges.size() - 1);
     }
 
     void RootSignatureGenerator::AddHeapRangesParameter(
